Change JoystickMovedListener.OnMoved() to use floats instead of ints

+++ src/com/MobileAnarchy/Android/Widgets/Joystick/JoystickMovedListener.java
@@ -1,7 +1,7 @@
 package com.MobileAnarchy.Android.Widgets.Joystick;
 
 public interface JoystickMovedListener {
-    public void OnMoved(int pan, int tilt);
+    public void OnMoved(float pan, float tilt);
     public void OnReleased();
     public void OnReturnedToCenter();
 }
